Use git, pull from github repo
