+++ pom.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.github.jmchilton.galaxybootstrap</groupId>
+  <groupId>com.github.dfornika.galaxybootstrap</groupId>
   <artifactId>galaxybootstrap</artifactId>
   <packaging>jar</packaging>
-  <version>0.6.1-SNAPSHOT</version>
+  <version>0.7.0-SNAPSHOT</version>
   <name>galaxy-bootstrap</name>
   <description />
   <parent>
@@ -13,13 +13,13 @@
     <version>7</version>
   </parent>
   <scm>
-    <connection>scm:git:git@github.com:jmchilton/galaxy-bootstrap.git</connection>
-    <developerConnection>scm:git:git@github.com:jmchilton/galaxy-bootstrap.git</developerConnection>
-    <url>git@github.com:jmchilton/galaxy-bootstrap.git</url>
+    <connection>scm:git:git@github.com:dfornika/galaxy-bootstrap.git</connection>
+    <developerConnection>scm:git:git@github.com:dfornika/galaxy-bootstrap.git</developerConnection>
+    <url>git@github.com:dfornika/galaxy-bootstrap.git</url>
   </scm>
   <issueManagement>
     <system>Github</system>
-    <url>http://github.com/jmchilton/galaxy-bootstrap/issues</url>
+    <url>http://github.com/dfornika/galaxy-bootstrap/issues</url>
   </issueManagement>
   <licenses>
     <license>
@@ -65,13 +65,36 @@
       <scope>test</scope>
     </dependency>
 
+    <!-- reporting -->
+    <dependency>
+      <groupId>org.apache.maven.plugins</groupId>
+      <artifactId>maven-project-info-reports-plugin</artifactId>
+      <version>2.9</version>
+      <type>maven-plugin</type>
+    </dependency>
+
+    <!-- documentation -->
+    <dependency>
+      <groupId>org.apache.maven.plugins</groupId>
+      <artifactId>maven-javadoc-plugin</artifactId>
+      <version>2.9</version>
+      <type>maven-plugin</type>
+    </dependency>
+
+    <!-- encryption -->
+    <dependency>
+      <groupId>org.apache.maven.plugins</groupId>
+      <artifactId>maven-gpg-plugin</artifactId>
+      <version>1.4</version>
+      <type>maven-plugin</type>
+    </dependency>
   </dependencies>
   <reporting>
     <plugins>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-project-info-reports-plugin</artifactId>
-        <version>2.7</version>
+        <version>2.9</version>
         <reportSets>
           <reportSet>
             <reports>
