+++ src/main/java/com/github/dfornika/galaxybootstrap/Config.java
@@ -1,4 +1,4 @@
-package com.github.jmchilton.galaxybootstrap;
+package com.github.dfornika.galaxybootstrap;
 
 import java.io.File;
