+++ src/main/resources/com/github/dfornika/galaxybootstrap/shed_tool_conf.xml
