+++ src/main/resources/com/github/dfornika/galaxybootstrap/universe.sqlite
